Fix bad quality of overlay image when...
When a video is in a bad quality and much less than an overlay.
In this case it would be much better to keep an overlay the original size and scale a bad quality video up even though it's still bad quality. This change allows not scaling an overlay down to the bad quality video.